requirements.txt
pandas==2.2.3
pyspark==3.5.1
apache-airflow

